documentcloud/documents/processing/utils/requirements.txt
-r cloud-requirements.txt

django-environ==0.4.5
requests==2.22.0
redis==3.2.1
furl==2.1.0
